+++ toolkit/views/account/confirm_account_creation.py
