+++ toolkit/views/account/logout.py
