+++ toolkit/views/saved/saved_loot.py
